Refactor Components to avoid Prop Drilling
